requirements.txt
uvicorn
fastapi
orjson
python-dotenv
aiohttp
pymysql
cryptography
mysqlclient